oopsie

+++ app.py
@@ -73,7 +73,7 @@ def collage(image_paths):
             y += single_height
         x += single_width+divider
         y = 0
-    out.paste(versus,((single_width*cols+divider)-88/2, (single_height*rows)-75/2), versus)
+    out.paste(versus,(int((single_width*cols+divider)-88/2), int((single_height*rows)-75/2)), versus)
     out.save("Collage.jpg")
